+++ tests/testthat/test-locales.R
@@ -8,6 +8,7 @@ testthat::test_that("Check locales are equal to ms_locales", {
                  "speech-service/language-support")
     doc = xml2::read_html(url)
     tab = rvest::html_table(doc)
+    # standard voices not neural
     index = sapply(tab, function(x) {
       if (!"Voice name" %in% colnames(x)) {
         return(FALSE)
@@ -20,6 +21,7 @@ testthat::test_that("Check locales are equal to ms_locales", {
 
     sub_tab = tab[, c("Locale", "Language")]
     sub_tab$Locale = gsub("[*]", "",     sub_tab$Locale)
+    sub_tab$Language = gsub(" \\)", ")",   sub_tab$Language)
     sub_tab = unique(sub_tab)
     sub_tab = sub_tab[!(sub_tab$Locale == "" & sub_tab$Language == ""),]
     cn = colnames(sub_tab)